fix: move open handling to web-contents.js
